refactor: dedupe fragments, move to local fragments, queries and types
