+++ front/src/hooks/useAssignBoxesToShipment.ts
@@ -4,7 +4,7 @@ import { useCallback, useState } from "react";
 import { IBoxBasicFields } from "types/graphql-local-only";
 import { useErrorHandling } from "./useErrorHandling";
 import { useNotification } from "./useNotification";
-import { SHIPMENT_FIELDS_FRAGMENT } from "../../../graphql/fragments";
+import { SHIPMENT_FIELDS_FRAGMENT } from "queries/fragments";
 
 export enum IAssignBoxToShipmentResultKind {
   BAD_USER_INPUT = "badUserInput", // no Boxes InStock were passed to the function
