+++ front/src/queries/fragments.ts
@@ -1,4 +1,4 @@
-import { BASE_BASIC_FIELDS_FRAGMENT, ORGANISATION_BASIC_FIELDS_FRAGMENT, PRODUCT_BASIC_FIELDS_FRAGMENT, SIZE_BASIC_FIELDS_FRAGMENT, TAG_BASIC_FIELDS_FRAGMENT, USER_BASIC_FIELDS_FRAGMENT } from "../../../graphql/fragments";
+import { BASE_BASIC_FIELDS_FRAGMENT, BOX_BASIC_FIELDS_FRAGMENT, ORGANISATION_BASIC_FIELDS_FRAGMENT, PRODUCT_BASIC_FIELDS_FRAGMENT, SIZE_BASIC_FIELDS_FRAGMENT, SIZE_RANGE_FIELDS_FRAGMENT, USER_BASIC_FIELDS_FRAGMENT } from "../../../graphql/fragments";
 import { graphql } from "../../../graphql/graphql"
 
 export const LOCATION_BASIC_FIELDS_FRAGMENT = graphql(`
@@ -15,6 +15,16 @@ export const TAG_OPTIONS_FRAGMENT = graphql(`
     value: id
     label: name
     color
+  }
+`);
+
+export const TAG_BASIC_FIELDS_FRAGMENT = graphql(`
+  fragment TagBasicFields on Tag @_unmask {
+    id
+    name
+    color
+    description
+    type
   }
 `);
 
@@ -41,48 +51,6 @@ export const BASE_ORG_FIELDS_FRAGMENT = graphql(`
   }
 `, [ORGANISATION_BASIC_FIELDS_FRAGMENT, BASE_BASIC_FIELDS_FRAGMENT]);
 
-export const HISTORY_FIELDS_FRAGMENT = graphql(`
-  fragment HistoryFields on HistoryEntry @_unmask {
-    id
-    changes
-    changeDate
-    user {
-      ...UserBasicFields
-    }
-  }
-`, [USER_BASIC_FIELDS_FRAGMENT]);
-
-export const SIZE_RANGE_FIELDS_FRAGMENT = graphql(`
-  fragment SizeRangeFields on SizeRange @_unmask {
-    id
-    label
-    sizes {
-      ...SizeBasicFields
-    }
-  }
-`, [SIZE_BASIC_FIELDS_FRAGMENT]);
-
-export const BOX_BASIC_FIELDS_FRAGMENT = graphql(`
-  fragment BoxBasicFields on Box @_unmask {
-    labelIdentifier
-    state
-    comment
-    location {
-      id
-      base {
-        id
-      }
-    }
-    shipmentDetail {
-      id
-      shipment {
-        id
-      }
-    }
-    lastModifiedOn
-  }
-`);
-
 // complexer fragments
 export const PRODUCT_FIELDS_FRAGMENT = graphql(`
   fragment ProductFields on Product @_unmask {
@@ -95,6 +63,18 @@ export const PRODUCT_FIELDS_FRAGMENT = graphql(`
     }
   }
 `, [PRODUCT_BASIC_FIELDS_FRAGMENT, SIZE_RANGE_FIELDS_FRAGMENT]);
+
+export const HISTORY_FIELDS_FRAGMENT = graphql(`
+  fragment HistoryFields on HistoryEntry @_unmask {
+    id
+    changes
+    changeDate
+    user {
+      ...UserBasicFields
+    }
+  }
+`, [USER_BASIC_FIELDS_FRAGMENT]);
+
 
 
 export const BOX_FIELDS_FRAGMENT = graphql(`
@@ -169,3 +149,136 @@ export const BOX_FIELDS_FRAGMENT = graphql(`
     lastModifiedOn
   }
 `, [PRODUCT_BASIC_FIELDS_FRAGMENT, SIZE_BASIC_FIELDS_FRAGMENT, TAG_BASIC_FIELDS_FRAGMENT, BASE_BASIC_FIELDS_FRAGMENT, HISTORY_FIELDS_FRAGMENT, LOCATION_BASIC_FIELDS_FRAGMENT]);
+
+export const SHIPMENT_DETAIL_FIELDS_FRAGMENT = graphql(`
+  fragment ShipmentDetailFields on ShipmentDetail @_unmask {
+    id
+    box {
+      ...BoxBasicFields
+    }
+    sourceProduct {
+      ...ProductFields
+    }
+    targetProduct {
+      ...ProductFields
+    }
+    sourceSize {
+      ...SizeBasicFields
+    }
+    targetSize {
+      ...SizeBasicFields
+    }
+    sourceLocation {
+      id
+      name
+      ... on ClassicLocation {
+        defaultBoxState
+      }
+    }
+    sourceQuantity
+    targetQuantity
+    createdOn
+    createdBy {
+      ...UserBasicFields
+    }
+    removedOn
+    removedBy {
+      ...UserBasicFields
+    }
+    lostOn
+    lostBy {
+      ...UserBasicFields
+    }
+    receivedOn
+    receivedBy {
+      ...UserBasicFields
+    }
+  }
+`, [BOX_BASIC_FIELDS_FRAGMENT, USER_BASIC_FIELDS_FRAGMENT, PRODUCT_FIELDS_FRAGMENT, SIZE_BASIC_FIELDS_FRAGMENT]);
+
+export const SHIPMENT_FIELDS_FRAGMENT = graphql(`
+  fragment ShipmentFields on Shipment @_unmask {
+    id
+    labelIdentifier
+    state
+    details {
+      ...ShipmentDetailFields
+    }
+    sourceBase {
+      ...BaseOrgFields
+    }
+    targetBase {
+      ...BaseOrgFields
+    }
+    transferAgreement {
+      id
+      comment
+      type
+    }
+    startedOn
+    startedBy {
+      ...UserBasicFields
+    }
+    sentOn
+    sentBy {
+      ...UserBasicFields
+    }
+    receivingStartedOn
+    receivingStartedBy {
+      ...UserBasicFields
+    }
+    completedOn
+    completedBy {
+      ...UserBasicFields
+    }
+    canceledOn
+    canceledBy {
+      ...UserBasicFields
+    }
+  }
+`, [BASE_ORG_FIELDS_FRAGMENT, SHIPMENT_DETAIL_FIELDS_FRAGMENT, USER_BASIC_FIELDS_FRAGMENT]);
+
+export const TRANSFER_AGREEMENT_FIELDS_FRAGMENT = graphql(`
+  fragment TransferAgreementFields on TransferAgreement @_unmask {
+    id
+    type
+    state
+    comment
+    validFrom
+    validUntil
+    sourceOrganisation {
+      ...OrganisationBasicFields
+    }
+    sourceBases {
+      ...BaseBasicFields
+    }
+    targetOrganisation {
+      ...OrganisationBasicFields
+    }
+    targetBases {
+      ...BaseBasicFields
+    }
+    shipments {
+      id
+      state
+      sourceBase {
+        ...BaseBasicFields
+      }
+      targetBase {
+        ...BaseBasicFields
+      }
+    }
+    requestedOn
+    requestedBy {
+      ...UserBasicFields
+    }
+    acceptedOn
+    acceptedBy {
+      ...UserBasicFields
+    }
+    terminatedOn
+    terminatedBy {
+      ...UserBasicFields
+    }
+  }
+`, [ORGANISATION_BASIC_FIELDS_FRAGMENT, BASE_BASIC_FIELDS_FRAGMENT, USER_BASIC_FIELDS_FRAGMENT]);
