+++ front/src/views/Boxes/BoxesView.tsx
@@ -11,9 +11,8 @@ import { useBaseIdParam } from "hooks/useBaseIdParam";
 import {
   PRODUCT_BASIC_FIELDS_FRAGMENT,
   SIZE_BASIC_FIELDS_FRAGMENT,
-  TAG_BASIC_FIELDS_FRAGMENT,
 } from "../../../../graphql/fragments";
-import { BASE_ORG_FIELDS_FRAGMENT } from "queries/fragments";
+import { BASE_ORG_FIELDS_FRAGMENT, TAG_BASIC_FIELDS_FRAGMENT } from "queries/fragments";
 import { BoxRow } from "./components/types";
 import { SelectColumnFilter } from "components/Table/Filter";
 import BoxesActionsAndTable from "./components/BoxesActionsAndTable";
