+++ front/src/views/Transfers/CreateTransferAgreement/CreateTransferAgreementView.tsx
@@ -15,7 +15,7 @@ import CreateTransferAgreement, {
 import { ALL_ACCEPTED_TRANSFER_AGREEMENTS_QUERY } from "../CreateShipment/CreateShipmentView";
 import { useBaseIdParam } from "hooks/useBaseIdParam";
 import { useLoadAndSetGlobalPreferences } from "hooks/useLoadAndSetGlobalPreferences";
-import { TRANSFER_AGREEMENT_FIELDS_FRAGMENT } from "../../../../../graphql/fragments";
+import { TRANSFER_AGREEMENT_FIELDS_FRAGMENT } from "queries/fragments";
 
 export const ALL_ORGS_AND_BASES_QUERY = graphql(`
   query AllOrganisationsAndBases {
